+++ src/remastered/color.js
@@ -1,11 +1,12 @@
 module.exports = {
   white: '#ffffff',
   lightGray: '#c9d6da',
-  darkGray: '#617d8a', // not used yet
+  darkGray: '#617d8a',
   gray: '#abb2bf',
-  purple: '#a39cf9', // not used yet
-  blue: '#6baaf8',
-  sky: '#73bae5', // not used yet
+  purple: '#b39afd',
+  blue: '#64B5F6',
+  sky: '#87cbff',
+  cyan: '#4DD0E1',
   green: '#3ca29c',
   teal: '#82c781',
   yellow: '#e5c275',